Change colors. Now the id names don't make sense.
If the client wants to change the red to purple and the green to orange we're now stuck because the id names are named after the color.

If the site is huge we might be stuck with those id names if we didn't use something like SASS.

Bootstrap gets around this scenario by using differnt words instead of color names.

- Primary
- Secondary
- Success
- Danger
- Warning
- Info
- Dark
- White
- Transparent